# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Project - OpenAPI 3.0
  version: 1.0.0
servers:
  - url: http://localhost:8081
tags:
  - name: Customer
    description: Operations about customer
paths:
  /customer:
    get:
      tags:
        - Customer
      summary: Get customer all
      description: CustomerEntity list from mongoDB.
      operationId: getCustomer
      responses:
        200:
          description: success get
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
    post:
      tags:
        - Customer
      summary: Create customer
      description: Creation of customer in mongoDB database.
      operationId: createCustomer
      requestBody:
        description: Created customer object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
        required: true
      responses:
        201:
          description: success created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
  /customer/{clientId}:
    get:
      tags:
        - Customer
      summary: Get customer by clientId
      description: Get customer.
      operationId: getCustomerById
      parameters:
        - name: clientId
          in: path
          description: find by clientId
          required: true
          schema:
            type: string
      responses:
        200:
          description: success get
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
    put:
      tags:
        - Customer
      summary: Update customer by clientId
      description: Update customer.
      operationId: updateCustomerById
      parameters:
        - name: clientId
          in: path
          description: id that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent user in the mongoDB
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
        required: true
      responses:
        200:
          description: success updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
    delete:
      tags:
        - Customer
      summary: Delete customer by clientId
      description: Delete an existent customer in the database mongoDB.
      operationId: deleteCustomerById
      parameters:
        - name: clientId
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        default:
          description: success delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDTO'
        404:
          description: CustomerEntity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDTO'
  /customer/document/{number}:
    get:
      tags:
        - Customer
      summary: Get customer by documentNumber
      description: Get customer.
      operationId: getCustomerByDocumentNumber
      parameters:
        - name: number
          in: path
          description: find by documentNumber
          required: true
          schema:
            type: string
      responses:
        200:
          description: success get
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
components:
  schemas:
    CustomerRequest:
      required:
        - name
        - lastName
        - documentType
        - documentNumber
        - clientType
      type: object
      properties:
        name:
          type: string
          example: John
          #pattern: '^[A-Za-z]+$'
        lastName:
          type: string
          example: James
          #pattern: '^[A-Za-z]+$'
        reason:
          type: string
          example: 'EMPRESA TRANSPORTE SAC'
        documentType:
          type: string
          enum:
            - DNI
            - RUC
          example: DNI
        documentNumber:
          type: string
          example: '45871236'
          pattern: '^[0-9]+$'
        clientType:
          type: string
          enum:
            - STAFF
            - BUSINESS
          example: STAFF
        address:
          maxLength: 20
          type: string
          example: "Av. Siempre viva #345"
        email:
          type: string
          format: email
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
          example: "usuario@ejemplo.com"
        phones:
          type: array
          items:
            type: string
            pattern: '^\+?[1-9]\d{1,14}$'
          example:
            - "+51987654321"
            - "+54976453245"
            - "+189764532"
    CustomerResponse:
      type: object
      properties:
        id:
          type: string
          example: '6680e3e7878a8f7d4d998569'
        name:
          type: string
          example: John
        lastName:
          type: string
          example: James
        reason:
          type: string
          example: 'EMPRESA TRANSPORTE SAC'
        documentType:
          type: string
          enum:
            - DNI
            - RUC
          example: DNI
        documentNumber:
          type: string
          example: '45871236'
        clientType:
          type: string
          enum:
            - STAFF
            - BUSINESS
          example: STAFF
        address:
          maxLength: 20
          type: string
          example: "Av. Siempre viva #34"
        email:
          type: string
          format: email
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
          example: "usuario@ejemplo.com"
        phones:
          type: array
          items:
            type: string
            pattern: '^\+?[1-9]\d{1,14}$'
          example:
            - "+51987654321"
            - "+54976453245"
            - "+189764532"
    ResponseDTO:
      type: object
      properties:
        message:
          type: string
          example: 'Message response'